добавлен хук изменения состояния инпута
